BE/scripts/down-sampling/requirements.txt
boto3==1.33.2
botocore==1.33.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
idna==3.6
jmespath==1.0.1
Pillow==10.1.0
pycparser==2.21
pyheif==0.7.1
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.8.1
six==1.16.0
urllib3==1.26.18
